+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## - 2021-07-23
+### Added
+- Added pipe checker.
+
 ## - 2021-07-22
 ### Added
 - Added pipe feature.
